[@azure/arm-newrelicobservability] Fix the Test Case (#30021)
### Packages impacted by this PR

@azure/arm-newrelicobservability

### Issues associated with this PR

None

### Describe the problem that is addressed by this PR

On June 6, 2024, an issue was reported by Manik Khandelwal as the `rushx
build` command is failing for the `@azure/arm-newrelicobservability`
SDK. The following error is reported:

```
test/newrelicobservability_operations_test.spec.ts.ts(56,7): error TS2353: Object literal may only specify known properties, and 'clientId' does not exist in type 'CreateTestCredentialOptions'.
test/newrelicobservability_operations_test.spec.ts.ts(60,56): error TS2554: Expected 0-1 arguments, but got 2.
```

This PR has been created to fix this issue. 

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

Looking at the root cause of the issue, the current code changes in the
`newrelicobservability_operations_test.spec.ts.ts` (with client id &
client secret) are added on March 21, 2024 (Refer
https://github.com/Azure/azure-sdk-for-js/pull/28829). At that time,
there was no problem with the code and it was working fine.

On May 6, 2024 some changes have been done in the
`@azure-tools/test-credential` in which the `clientId` and
`clientSecret` have been removed. Thus the reported error is happening.
(Now, Ideally I would have expected that such issues would have been
flagged by our CI while merging the changes for
`@azure-tools/test-credential`. But, that did not happen. Refer the CI
run in https://github.com/Azure/azure-sdk-for-js/pull/29577.
@HarshaNalluru for further investigation)

In order to resolve the issue, I have removed the extra parameters. The
`playback` mode is also running fine. The CI could be seen at
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3868221&view=results.
Because of the artifact already exists, the check enforcer is not
completing. But that is not a serious issue. I am overriding the check
enforcer.

### Are there test cases added in this PR? _(If not, why?)_

No. The issue itself is in the test case.

### Provide a list of related PRs _(if any)_

1. https://github.com/Azure/azure-sdk-for-js/pull/28829
2. https://github.com/Azure/azure-sdk-for-js/pull/29577

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@xirzec @joheredi Please review and approve the PR