Update the Core Client Version and Test Recorder (#30032)
### Packages impacted by this PR
@Azure/communication-chat 

### Issues associated with this PR
### Describe the problem that is addressed by this PR

Customer escalated the issue:
The issue was introduced from the @azure/communication-chat@1.4.0. Chat
client updated the internal API client by generating the new service
client version with the `endpoint` param to replace the `baseUri` and
kept the dependency version @azure/core-client@1.3.0 and higher. While
the `endpoint` is introduced only in @azure/core-client@1.6.0.
Customer see error "baseUrl is not specified" if the dependency version
not updated.

Update the @azure/core-client version to fix the issue


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)