[@azure/monitor-query] Live Tests CI Failure (#30081)
### Packages impacted by this PR

@azure/monitor-query

### Issues associated with this PR

None

### Describe the problem that is addressed by this PR

The live tests for `@azure/monitor-query` are failing daily for only one
configuration `ubuntu2004_node_18x_min`. Example:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3864316&view=logs&j=0daa86f4-d78e-5a1c-68e8-8dcfd7021055.
The error message is the following:

```
Invoking "rushx build"
----------------------

Found configuration in /mnt/vss/_work/1/s/rush.json

Rush Multi-Project Build Tool 5.120.6 - Node.js 18.20.3 (LTS)
> "tsc -p ."

../../../../../common/temp/node_modules/.pnpm/@azure-tools+test-recorder@3.5.0/node_modules/@azure-tools/test-recorder/types/src/recorder.d.ts(6,10): error TS2305: Module '"@azure/core-client"' has no exported member 'AdditionalPolicyConfig'.
Error: Failed calling tsc -p ..  Exit code: 2
```

In order to fix this issue, we need to update the version of the
dependency `@azure/core-client` to `1.9.2`. This PR does that and fixes
the problem.

I have executed the live tests CI with the changes in this PR at
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3879546&view=results.
The results are successful.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No special design concerns.

### Are there test cases added in this PR? _(If not, why?)_

No code changes are involved. The existing test cases are sufficient for
the configuration change.

### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/29931

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@HarshaNalluru @minhanh-phan @jeremymeng Please review and approve the
PR.

@xirzec FYI......