[core] Provide `error` parameter to onResponse callback (#29983)
### Packages impacted by this PR

- `@azure-rest/core-client`

### Describe the problem that is addressed by this PR

Proposal to mitigate a couple of breaking changes between the old and
new `core-client` packages:
- `onResponse` is now called when an error is thrown while sending the
request. This was not being done by `core-client-rest` but [is being
done by `core-client`

](https://github.com/Azure/azure-sdk-for-js/blob/9f03724468521497f8499907433a715b062a9884/sdk/core/core-client/src/serviceClient.ts#L217-L229).
- Updated the type of the `onResponse` callback to be more
backwards-compatible with the old Core. The old Core's callback had
three parameters (response, flatResponse, and error), but the new
package has only two (response and error). The `flatResponse` property
does not make sense in the new Core and we have decided not to support
this property. In order to enable this, we pass the error twice in the
new callback so that we can support passing the `error` through to both
the old and new variations of `onResponse`.