[Eng] Update Env Var for Execute Samples Task (#30132)
### Describe the problem that is addressed by this PR
Our samples currently use `DefaultAzureCredential` to authenticate the
client. This causes a problem running samples live in the pipeline
because Managed Identity does not work in the live pipeline. This is a
short term work around: we create a fake authority host so that Managed
Identity can fail fast and it will succeed with the Powershell credential.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
We do not want to change the samples to use a different credentials
because we want our customers to use DAC. We are discussing a long-term
solution by potentially adding exclude managed identity property similar
to Python/ .NET