[CODEOWNERS] Avoid bottleneck in CODEOWNERS (#30188)
### Describe the problem that is addressed by this PR

The issue tracking system works in a way that more specific paths
override more general ones. This means that a rule like `/sdk/ @xirzec`
will be overridden by a more specific rule like `/sdk/foo @someOwner`
when assigning issues.

~~However, when considering required reviewers based on CODEOWNERS,
GitHub does *not* take this approach, leading to the catch-all person
needing to approve *all* reviews that touch that path.~~

This creates a problem when introducing a new service folder as nobody
is assigned to approve it. To avoid myself becoming the bottleneck in
such cases, we can add the group to this line as it does not have
notifications enabled. I also added more approvers for `rush.json` which
is affected by new package generation.