[Cosmos] Split/Merge support for Bulk API (#28725)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
[Split Proof Bulk / Batch / Patch APIs in Cosmos JS SDK · Issue #18682 ·
Azure/azure-sdk-for-js
(github.com)](https://github.com/Azure/azure-sdk-for-js/issues/18682)
At present, if the bulk API SDK encounters a partition merge or split
during the process, it returns an error. Please review the process
outlined below:


![image](https://github.com/Azure/azure-sdk-for-js/assets/113669638/99a15e72-b019-48e5-bc02-d9f064e6309d)


### Describe the problem that is addressed by this PR
This PR ensures that the Bulk API is resistant to partition splitting
and merging.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
We've implemented a BFS (breadth-first search) based approach. This
method involves splitting batch items according to the new partition
range and adding them back to the FIFO queue(newly introduced) to manage
partition splits/merges. Please refer to the following process for more
details.

![image](https://github.com/Azure/azure-sdk-for-js/assets/113669638/33648faa-d06d-47b0-b017-4097bed56b92)


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Manik Khandelwal <mkhandelwal@microsoft.com>