[core] Make `request` and `response` properties on RestError non-enumerable (#30518)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`
- `@typespec/ts-http-runtime`

### Issues associated with this PR

- See e.g. https://github.com/Azure/azure-sdk-for-js/issues/30247
- Also https://github.com/Azure/azure-sdk-for-js/issues/29910
- And #29630

### Describe the problem that is addressed by this PR

Taken from my last PR #30483:

> `PipelineRequest` and `PipelineResponse` objects may contain secrets
in the request URL or headers. This is problematic since `RestError`
objects are often logged. While we override `util.inspect.custom` to
sanitize log output in Node, this does not work in all situations and
environments. For example:
> - `console.log` in browser does not respect `util.inspect.custom`,
meaning secrets could be logged to the browser console. Other non-Node
environments also may not respect this Node-specific functionality.
> - JSON serialization of `RestError` objects. Calling `JSON.stringify`
on the `RestError` currently results in an object that contains
unsanitized secrets. We have encountered scenarios where the JSON
serialization is logged, for example with vitest.

This PR fixes this issue by making the `request` and `response`
properties **non-enumerable**. This means they are ignored by
JSON.stringify and `Object.entries`/`Object.keys`, but the properties
are still directly accessible. This should prevent any potential
sensitive information being accidentally logged in the majority of
scenarios. I think this is a better balance of not breaking folks versus
improving security than the approach in #30483.