[@azure/monitor-query] Fixed the audience values for MetricsQueryClient and MetricsClient (#30589)
### Packages impacted by this PR
@azure/monitor-query

### Issues associated with this PR
NA

### Describe the problem that is addressed by this PR
The audience values set for MetricsQueryClient is incorrect and has to
be fixed. Also, new Audience values are introduced for LogsQueryClient
and MetricsClient.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
No specific design considerations or big API changes.

### Are there test cases added in this PR? _(If not, why?)_
No. They are just changes to enum values.

### Provide a list of related PRs _(if any)_
NA

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
NA

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)


@srnagar @xirzec Please review and approve the PR.