[@azure/cosmos] Add check for TOP/LIMIT in nonStreamingOrderBy queries (#30503)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR


### Describe the problem that is addressed by this PR
- This PR removes the support of float 16 from vector embedding data
type
- This PR adds support to block `nonStreamingOrderBy` queries which do
not have a TOP or LIMIT+OFFSET present in the query. This check can be
disabled by using `allowUnboundedNonStreamingQueries` flag in options.

- An additional check is placed for internal buffer size. If calculated
buffer size is greater than the default buffer size, an error is thrown.
This is done as a safeguard from excessive memory consumption. Max
buffer size can be increased using `vectorSearchBufferSize` in the
options.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)