[DataLake] Update storage-blob includes in package.json (#30612)
Fix https://github.com/Azure/azure-sdk-for-js/issues/30463 by including
the portions of @azure/storage-blob required by
@azure/storage-file-datalake.


### Packages impacted by this PR

@azure/storage-file-datalake

### Issues associated with this PR

https://github.com/Azure/azure-sdk-for-js/issues/30463

### Describe the problem that is addressed by this PR

Imports are broken. The package is unusable.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

N/A

### Are there test cases added in this PR? _(If not, why?)_

No. No code has changed.

### Provide a list of related PRs _(if any)_

N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

N/A

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)