[core-sse] Fix NodeJS socket behavior (#30707)
### Packages impacted by this PR
@azure/core-sse

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/30414

### Describe the problem that is addressed by this PR
In NodeJS, when the response stream is exhausted and the server closes
the connection, the underlying socket gets set to null. This is
problematic because core-sse is misled by the type of
`http.IncomingMessage.socket` that stipulates the socket is always a
valid object. This PR checks first if the socket is defined before it
closes it.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
Yes!

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)