[eslint] fix ci failures (#30765)
Recently linting was enabled for `monitor-opentelemetry-exporter`. Its
eslint
config enables typed linting which requires its dependencies to be build
in
order to have proper types.

This PR passes `-t @azure/monitor-opentelemetry-exporter` to `rush
build`
before linting so that its dependencies are built.

While at it, also change the `run-for-all-packages` pipeline to use
`install-run-rush.js` for consistency.