[Event Hubs] [Bicep] Factor out KV definition (#30894)
Moving to a modular bicep definition so that we can share bits and
pieces between all of them. This PR factors out the key vault definition
so other bicep scripts can import it too. For now, it lives next to the
event hub's test-resources.bicep script but I'll consider moving it to a
central place once other libraries move to import it.

The PR also edits the script to make it more compact.

Live tests run:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4079360&view=results