[core-xml] fix stringifyXML missing XML namespace issue on Firefox
It looks that chrome-based browsers are more error-tolerant when the XML
namespace ("xmlns:...") was set via attribute, but that doesn't work on Firefox,
which requires using of `createElementNS()` to create element with XML
namespaces.

This PR adds special handling of the "xmlns" attribute. Tests passed so I didn't
do any additional work to remove the already processed "xmlns" from the
attribute list.