[test-credential] Include the correct types in package (#31039)
### Packages impacted by this PR

@azure-tools/test-credential

### Issues associated with this PR

N/A - hotfix 1.3.0

### Describe the problem that is addressed by this PR

In
https://github.com/Azure/azure-sdk-for-js/pull/30977/files#diff-f69713a41b4518b09d100824cfa2dc13e3105ab66b0f81f12a8290df980f441dL8
we updated the filename for the rollup dts file but forgot to update the
`files` section to ensure the new file is included in the published
package.

This hotfix focuses on the straightforward fix which is updating the
`files` declaration. Once this is released, I will also look at
opportunities for us to catch this before publishing the package. I
don't know what that will look like yet, but for now let's fix the
published package