[core] Create spans using automatic span propagation (#31019)
### Packages impacted by this PR

@azure/core-rest-pipeline
@typespec/ts-http-runtime

### Issues associated with this PR

N/A

### Describe the problem that is addressed by this PR

Back when we were migrating to the new core tracing we wanted to be
cautious and
bail as early as possible in our core tracing policy. This meant that
we:
1. Guard against any errors
2. Ignore automatic span propagation, instead expecting HLC libraries to
call tracingClient.withSpan to kick off the manual span propagation (context
parameter passing in options)

Now that we have RLC which does not have a tracingClient, and having had
this code in production for quite some time, we can remove the optimization
that aassumes the tracingContext will be added upstream.

This is a non-impactful change for high-level clients or anyone who has
not opted-in to OTel tracing.

For users who use rest-level clients _and_ have opted into OTel, they
will now see HTTP spans from core added correctly.