[keyvault-certificates] fix minmax failure (#31084)
### Packages impacted by this PR

- `@azure/keyvault-certificates`

### Issues associated with this PR

- Fix #31012

### Describe the problem that is addressed by this PR

1. Minmax tests are not run in the package directory (they are instead
run in `test/public`). This PR updates the path for some static test
resource files to be relative to the test file itself instead of
relative to the working directory.
2. Stop using instanceof for a couple of tests where we are testing the
type of the error. The name of the error should be enough. I think this
might have been causing failures maybe due to a dual-package hazard kind
of interaction?