[tracing] Add support for addEvent (#31162)
### Packages impacted by this PR

@azure/core-tracing
@azure/opentelemetry-instrumentation-azure-sdk
@typespec/ts-http-runtime

### Issues associated with this PR

Resolves #31158

### Describe the problem that is addressed by this PR

OpenTelemetry supports adding events to a span, and part of the gen-ai
specification expects events to be added on a span. 

This PR adds support for the `addEvent` API in our core-tracing library
as well as our otel-sdk.

Adding a required field is a breaking change, so this is added as
optional. The implementation will need to use the safe-navigation 
operator syntax or check for the function's existence before calling it.

In passing, I am updating some of our internal transformations to use
OTel's APIs instead.