[@azure/eventgrid] Update System events for EG Version 5.7.0  (#31186)
### Packages impacted by this PR

@azure/eventgrid

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

This PR consists of the following changes:

- A new property `tierToColdSummary` is added to the
`StorageLifecyclePolicyCompletedEventData` interface.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

There are no specific/complex design scenarios for this task. It is a
straightforward regenerate and some standard changes to the custom layer
of the code.

### Are there test cases added in this PR? _(If not, why?)_
No. This item is standard and we need not add test cases for every new
events. The existing cases would be sufficient.

### Provide a list of related PRs _(if any)_
- https://github.com/Azure/azure-sdk-for-js/pull/28176 (This is the PR
that adds similar events to the SDK in the 5.1.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/28513 (This is the PR
that adds similar events to the SDK in the 5.2.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/28891 (This is the PR
that adds similar events to the SDK in the 5.3.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/29035 (This is the PR
that adds similar events to the SDK in the 5.4.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/29872 (This is the PR
that adds similar events to the SDK in the 5.5.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/30811 (This is the PR
that adds similar events to the SDK in the 5.6.0 release)

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
```autorest --typescript swagger\README.md```

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so, create an Issue in the [Autorest/typescript](https://github.com/Azure/autorest.typescript) repository and link it here)_
- [X] Added a changelog (if necessary)