[dev-tool] Add build-package command (#31197)
### Packages impacted by this PR

core

### Issues associated with this PR

N/A

### Describe the problem that is addressed by this PR

As we look to migrate all packages to tshy we want to make sure we are
able to move away from tshy if ever needed. This PR adds a build-package command
that can later be used to consolidate some of the commands and decouple a bit
from tshy

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

We can just call tshy directly, but one of the "selling" points of tshy
is that we can always eject and do what it does manually if we run into issues.
This helps streamline that without having to update 400 package.json files if
needed.

Frankly, I am not 100% sure if this is needed and I admit that `tshy` is
still an entry in packages' json, but I think we can live with the `tshy`
entry longer than we can with a `tshy` tool migration.