[core-client-rest] Add support for `allowReserved` and `explode` (#31058)
### Packages impacted by this PR

- `@azure-rest/core-client`

### Issues associated with this PR

- Fixes #30943
- Fixes #30944

### Describe the problem that is addressed by this PR

Adds support for passing in an object as a query or path parameter. At
present, this object allows for the user to specify a couple of
encoding-related options in addition to the value of the parameter:
- For path parameters: `allowReserved` can be set to true to indicate
that special characters appearing in the parameter should not be
URL-encoded. This allows for path parameters to contain characters like
`/`.
- For query parameters: `explode` can be set to true to change how
arrays are encoded. By default, array values are encoded as a single
query parameter, with each value separated by a comma (e.g. `foo=a,b,c`.
With `explode` set to true, each array value will be converted into a
single query parameter (e.g. `foo=a&foo=b&foo=c`).