Howie/inference samples (#31316)
### Packages impacted by this PR
Nothing.  Sample only

### Issues associated with this PR
Fixed samples to use proper environment variables. Use latest versions
of core tracing and instrumentation sdk.
I don't want users to change code to make the image url to work. So
change the code to make it run out of the box.

### Describe the problem that is addressed by this PR


### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)