[ServiceBus] Add the experimental diagnostic feature `omitMessageBody`
via a module-level function `peekMessages()` under `./experimental` subpath export.

After migrate to ESM, we no longer use rolled-up dts files thus lose the ability
to utilize the pulic-trimmed and beta-trimmed rollups to exclude/include `@beta`
Apis in the type definition files. This PR explores an approach to expose beta
Apis from the `./experimental` export.

I found that maintaining a full copy of source files for the chain
`ServiceBusClient -> ServiceBusReceiver -> peekMessages -> peekMessagesOptions`
is too much for just an additional property in the option bag. So I am wrapping
the receiver with a module-level function instead. It is not the same as calling
client methods, but seems a close form as trade-off.