[core-client-rest] Fix serialization of non-string path parameters (#31352)
### Packages impacted by this PR

- `@azure-rest/core-client`

### Issues associated with this PR

- Fix https://github.com/Azure/azure-sdk-for-js/issues/31349

### Describe the problem that is addressed by this PR

We generate RLCs that allow for path parameters to be `number`. The
change for `allowReserved` assumed that path parameters could only be
`string`, causing number parameters to not be serialized properly. This
PR fixes that assumption.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_

Added a test to check for a number path parameter

### Provide a list of related PRs _(if any)_

- #31058