[engsys] Allow min/max tests to use vitest config of the target package (#31343)
### Packages impacted by this PR


### Issues associated with this PR

Resolves #30578 

### Describe the problem that is addressed by this PR

While simple settings like `test-timeout` and `hook-timeout` are often
defined in package.json scripts, there are
other settings that are specified in the vitest.config.ts of the target
package.

Our existing setup uses the default vitest template which does not work
for libraries that use customized config.

This PR improves that experience by creating a 3-way-merge of:
1. The base config
2. The template config (which specifies the test patterns)
3. The package's config

Only done for node right now. If you need it for browser, you now have a
path forward

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

There are some gotchas here:
1. testTimeout is defined in package.json which takes precedence over
any config files I believe - it may be surprising that your testTimeout
does not take (note that this is somewhat the case today)
2. The package's config adds files under `test/**/*.spec.ts` to the
include pattern. This is fine, because the min/max tests are placed
under `test/public` so that pattern will not match anything

I really want to clean up min/max - but this is not the time